Move localization files to more-logical places
