+++ src/webextension/l10n.js
@@ -9,16 +9,16 @@ import { LocalizationProvider } from "fluent-react";
 import PropTypes from "prop-types";
 import React, { Component } from "react";
 
-async function fetchMessages(locale) {
-  const response = await fetch(`${locale}.ftl`);
+async function fetchMessages(baseDir, locale) {
+  const response = await fetch(`${baseDir}/${locale}.ftl`);
   const messages = await response.text();
 
   return { [locale]: messages };
 }
 
-async function createMessagesGenerator(currentLocales) {
+async function createMessagesGenerator(baseDir, currentLocales) {
   const fetched = await Promise.all(
-    currentLocales.map(fetchMessages)
+    currentLocales.map((x) => fetchMessages(baseDir, x))
   );
   const bundle = fetched.reduce(
     (obj, cur) => Object.assign(obj, cur)
@@ -33,9 +33,10 @@ async function createMessagesGenerator(currentLocales) {
   }
 }
 
-export class AppLocalizationProvider extends Component {
+export default class AppLocalizationProvider extends Component {
   static get propTypes() {
     return {
+      baseDir: PropTypes.string,
       userLocales: PropTypes.array,
       children: PropTypes.any,
     };
@@ -44,7 +45,7 @@ export class AppLocalizationProvider extends Component {
   constructor(props) {
     super(props);
 
-    const { userLocales } = props;
+    const { baseDir = ".", userLocales } = props;
     const currentLocales = negotiateLanguages(
       userLocales, ["en-US"],
       { defaultLocale: "en-US" }
@@ -52,12 +53,15 @@ export class AppLocalizationProvider extends Component {
 
     this.state = {
       currentLocales,
+      baseDir,
     };
   }
 
   async componentWillMount() {
-    const { currentLocales } = this.state;
-    const generateMessages = await createMessagesGenerator(currentLocales);
+    const { baseDir, currentLocales } = this.state;
+    const generateMessages = await createMessagesGenerator(
+      baseDir, currentLocales
+    );
     this.setState({ messages: generateMessages() });
   }
 
@@ -66,7 +70,7 @@ export class AppLocalizationProvider extends Component {
     const { messages } = this.state;
 
     if (!messages) {
-      return <div>…</div>;
+      return <div/>;
     }
 
     return (
