+++ src/webextension/manage/locales/en-US.ftl
